# [Proposal] Make Performance test report HTML and to be collapsed by default